+++ Cargo.toml
@@ -3,7 +3,6 @@ resolver = "2"
 members = [
     "bin/roketz",
     "crates/bvh",
-    "crates/ecs",
 ]
 
 [profile.release]
@@ -11,7 +10,6 @@ strip = "symbols"
 
 [workspace.dependencies]
 bvh = { path = "crates/bvh" }
-ecs = { path = "crates/ecs" }
 
 anyhow = "1.0.98"
 bevy_ecs = "0.16.1"
