+++ bin/roketz/src/camera.rs
@@ -1,7 +1,7 @@
 use bevy_ecs::prelude::*;
 use macroquad::prelude::*;
 
-use ecs::cs::Transform;
+use crate::ecs::cs::Transform;
 
 pub enum CameraType {
     Global,
