+++ bin/roketz/src/ecs/cs/bullet.rs
@@ -3,7 +3,7 @@ use macroquad::prelude::*;
 use rapier2d::prelude::*;
 use tracing::error;
 
-use crate::{
+use crate::ecs::{
     cs::{Explosion, RigidCollider, TerrainCollider, Transform},
     r::{Collisions, DT, PhysicsWorld, Sound},
 };
