+++ bin/roketz/src/ecs/r/assets.rs
