--- crates/ecs/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "ecs"
-license = "MIT OR Apache-2.0"
-version = "0.0.0"
-edition = "2024"
-
-[features]
-default = []
-fmod = ["sound/fmod"]
-
-[dependencies]
-bvh.workspace = true
-sound = { path = "../sound", default-features = false, optional = true }
-
-anyhow.workspace = true
-bevy_ecs.workspace = true
-crossbeam.workspace = true
-knus.workspace = true
-macroquad.workspace = true
-rapier2d.workspace = true
-rdss.workspace = true
-tracing.workspace = true
